+++ x/fswap/handler.go
@@ -1,0 +1,26 @@
+package fswap
+
+import (
+	"fmt"
+
+	sdk "github.com/Finschia/finschia-sdk/types"
+	sdkerrors "github.com/Finschia/finschia-sdk/types/errors"
+	"github.com/Finschia/finschia-sdk/x/fswap/keeper"
+	"github.com/Finschia/finschia-sdk/x/fswap/types"
+)
+
+// NewHandler ...
+func NewHandler(k keeper.Keeper) sdk.Handler {
+	// this line is used by starport scaffolding # handler/msgServer
+
+	return func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
+		// ctx = ctx.WithEventManager(sdk.NewEventManager())
+
+		// switch msg := msg.(type) {
+		// this line is used by starport scaffolding # 1
+		// default:
+		errMsg := fmt.Sprintf("unrecognized %s message type: %T", types.ModuleName, msg)
+		return nil, sdkerrors.Wrap(sdkerrors.ErrUnknownRequest, errMsg)
+		// }
+	}
+}